refactor(UserService): use custom exception
